Rewrote entire project
Added to functions.py:
parse_pgn, move_to_idx, encode_move.

Added ChessDataset.py:
A way to make sure we retrieve data points in a format that is suitable for training.

Added ChessModel.py:
CNN for the model.

Added moves.py:
Generate the moves from a PGN file and save it to a separate file.

Added train_model.py and test_model.py:
A way to train and test the model.